WIP  Kore.Step.Simplification.Ceil: Remove evaluator call
