+++ kore/test/Test/Kore/Step/Simplification/AndTerms.hs
@@ -935,202 +935,6 @@ test_andTermsSimplification =
         ]
     ]
 
-test_equalsTermsSimplification :: [TestTree]
-test_equalsTermsSimplification =
-    [ testCase "adds ceil when producing substitutions" $ do
-        let expected = Just
-                [ Conditional
-                    { term = ()
-                    , predicate = makeCeilPredicate Mock.cf
-                    , substitution =
-                        Substitution.unsafeWrap [(ElemVar Mock.x, Mock.cf)]
-                    }
-                ]
-        actual <- simplifyEquals Map.empty (mkElemVar Mock.x) Mock.cf
-        assertEqualWithExplanation "" expected actual
-    , testCase "handles ambiguity" $ do
-        let
-            expected = Just
-                [ Conditional
-                    { term = ()
-                    , predicate = makeEqualsPredicate (Mock.f Mock.a) Mock.a
-                    , substitution =
-                        Substitution.unsafeWrap [(ElemVar Mock.x, Mock.cf)]
-                    }
-                , Conditional
-                    { term = ()
-                    , predicate = makeEqualsPredicate (Mock.f Mock.b) Mock.b
-                    , substitution =
-                        Substitution.unsafeWrap [(ElemVar Mock.x, Mock.cf)]
-                    }
-                ]
-            sortVar = SortVariableSort (SortVariable (testId "S"))
-            simplifiers = axiomPatternsToEvaluators $ Map.fromList
-                [   (   AxiomIdentifier.Ceil
-                            (AxiomIdentifier.Application Mock.cfId)
-                    ,   [ EqualityRule RulePattern
-                            { left = mkCeil sortVar Mock.cf
-                            , right =
-                                mkOr
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.f (mkElemVar Mock.y))
-                                            Mock.a
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.y) Mock.a)
-                                    )
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.f (mkElemVar Mock.y))
-                                            Mock.b
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.y) Mock.b)
-                                    )
-                            , requires = makeTruePredicate
-                            , ensures = makeTruePredicate
-                            , attributes = def
-                                {Attribute.simplification = Simplification True}
-                            }
-                        ]
-                    )
-                ]
-        actual <- simplifyEquals simplifiers (mkElemVar Mock.x) Mock.cf
-        assertEqualWithExplanation "" expected actual
-    , testCase "handles multiple ambiguity" $ do
-        let
-            expected = Just
-                [ Conditional
-                    { term = ()
-                    , predicate = makeAndPredicate
-                        (makeEqualsPredicate (Mock.f Mock.a) Mock.a)
-                        (makeEqualsPredicate (Mock.g Mock.a) Mock.a)
-                    , substitution = Substitution.unsafeWrap
-                        [ (ElemVar Mock.x, Mock.cf)
-                        , (ElemVar Mock.var_x_1, Mock.cg)
-                        ]
-                    }
-                , Conditional
-                    { term = ()
-                    , predicate = makeAndPredicate
-                        (makeEqualsPredicate (Mock.f Mock.a) Mock.a)
-                        (makeEqualsPredicate (Mock.g Mock.b) Mock.b)
-                    , substitution = Substitution.unsafeWrap
-                        [ (ElemVar Mock.x, Mock.cf)
-                        , (ElemVar Mock.var_x_1, Mock.cg)
-                        ]
-                    }
-                , Conditional
-                    { term = ()
-                    , predicate = makeAndPredicate
-                        (makeEqualsPredicate (Mock.f Mock.b) Mock.b)
-                        (makeEqualsPredicate (Mock.g Mock.a) Mock.a)
-                    , substitution = Substitution.unsafeWrap
-                        [ (ElemVar Mock.x, Mock.cf)
-                        , (ElemVar Mock.var_x_1, Mock.cg)
-                        ]
-                    }
-                , Conditional
-                    { term = ()
-                    , predicate = makeAndPredicate
-                        (makeEqualsPredicate (Mock.f Mock.b) Mock.b)
-                        (makeEqualsPredicate (Mock.g Mock.b) Mock.b)
-                    , substitution = Substitution.unsafeWrap
-                        [ (ElemVar Mock.x, Mock.cf)
-                        , (ElemVar Mock.var_x_1, Mock.cg)
-                        ]
-                    }
-                ]
-            sortVar = SortVariableSort (SortVariable (testId "S"))
-            simplifiers = axiomPatternsToEvaluators $ Map.fromList
-                [   (   AxiomIdentifier.Ceil
-                            (AxiomIdentifier.Application Mock.cfId)
-                    ,   [ EqualityRule RulePattern
-                            { left = mkCeil sortVar Mock.cf
-                            , right =
-                                mkOr
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.f (mkElemVar Mock.y))
-                                            Mock.a
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.y) Mock.a)
-                                    )
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.f (mkElemVar Mock.y))
-                                            Mock.b
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.y) Mock.b)
-                                    )
-                            , requires = makeTruePredicate
-                            , ensures = makeTruePredicate
-                            , attributes = def
-                                {Attribute.simplification = Simplification True}
-                            }
-                        ]
-                    )
-                ,   (   AxiomIdentifier.Ceil
-                            (AxiomIdentifier.Application Mock.cgId)
-                    ,   [ EqualityRule RulePattern
-                            { left = mkCeil sortVar Mock.cg
-                            , right =
-                                mkOr
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.g (mkElemVar Mock.z))
-                                            Mock.a
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.z) Mock.a)
-                                    )
-                                    (mkAnd
-                                        (mkEquals_
-                                            (Mock.g (mkElemVar Mock.z))
-                                            Mock.b
-                                        )
-                                        (mkEquals_ (mkElemVar Mock.z) Mock.b)
-                                    )
-                            , requires = makeTruePredicate
-                            , ensures = makeTruePredicate
-                            , attributes = def
-                                {Attribute.simplification = Simplification True}
-                            }
-                        ]
-                    )
-                ]
-        actual <- simplifyEquals
-            simplifiers
-            (Mock.functionalConstr20 (mkElemVar Mock.x) (mkElemVar Mock.var_x_1))
-            (Mock.functionalConstr20 Mock.cf Mock.cg)
-        assertEqualWithExplanation "" expected actual
-    , testCase "handles set ambiguity" $ do
-        let
-            asInternal set =
-                Ac.asInternalConcrete
-                    Mock.metadataTools
-                    Mock.setSort
-                    (Map.fromSet (const Domain.SetValue) set)
-            expected = Just $ do -- list monad
-                (xValue, xSetValue) <-
-                    [ (Mock.a, [Mock.b])
-                    , (Mock.b, [Mock.a])
-                    ]
-                return Conditional
-                    { term = ()
-                    , predicate = makeTruePredicate
-                    , substitution = Substitution.unsafeWrap
-                        [ (ElemVar Mock.x, xValue)
-                        , ( ElemVar Mock.xSet
-                          , asInternal (Set.fromList xSetValue)
-                          )
-                        ]
-                    }
-        actual <- simplifyEquals
-            Map.empty
-            (Mock.concatSet (Mock.elementSet (mkElemVar Mock.x)) (mkElemVar Mock.xSet))
-            (asInternal (Set.fromList [Mock.a, Mock.b]))
-        assertEqualWithExplanation "" expected actual
-    ]
-
 fOfA :: TermLike Variable
 fOfA = Mock.f Mock.a
